requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
appnope==0.1.3
arrow==1.2.2
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
base58==2.1.1
bitarray==2.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cloudinary==1.29.0
cycler==0.11.0
cytoolz==0.12.0
decorator==5.1.1
Deprecated==1.2.13
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
executing==0.10.0
Flask==2.2.2
Flask-Cors==3.0.10
fonttools==4.36.0
frozenlist==1.3.1
gunicorn==20.1.0
hexbytes==0.3.0
idna==3.3
importlib-metadata==4.12.0
ipfshttpclient==0.8.0a2
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.13.0
kiwisolver==1.4.4
lru-dict==1.1.8
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.5
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
numpy==1.23.2
packaging==21.3
parsimonious==0.8.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt==0.4.1
prompt-toolkit==3.0.30
protobuf==3.20.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycryptodome==3.15.0
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
redis==4.3.4
requests==2.28.1
rlp==2.0.1
rq==1.11.0
rq-dashboard==0.6.1
six==1.16.0
stack-data==0.4.0
toolz==0.12.0
tqdm==4.64.0
traitlets==5.3.0
urllib3==1.26.11
varint==1.0.2
wcwidth==0.2.5
web3==5.30.0
websockets==9.1
Werkzeug==2.2.2
wrapt==1.14.1
yarl==1.8.1
zipp==3.8.1
